try tofix err: 310

+++ src/api/challengeService.tsx
@@ -1,44 +1,70 @@
-import { API_BASE_URL, API_ENDPOINTS } from "@/constants/api";
+import { API_BASE_URL, API_ENDPOINTS, IS_DEV, IS_LOCAL } from "@/constants/api";
 
 export const challengeService = {
   async fetchChallenges() {
-    const response = await fetch(
-      `${API_BASE_URL}${API_ENDPOINTS.CHALLENGE.ALL}`
-    );
-    // const response = await fetch("/challanges.json");
-    if (!response.ok) {
-      throw new Error(`HTTP error! status: ${response.status}`);
-    }
 
-    return response.json();
+    if (IS_LOCAL && IS_DEV) {
+      const response = await fetch("/data/challanges.json");
+      if (!response.ok) {
+        throw new Error(`HTTP error! status: ${response.status}`);
+      }
+
+      return response.json();
+    } else {
+
+      const response = await fetch(
+        `${API_BASE_URL}${API_ENDPOINTS.CHALLENGE.ALL}`
+      );
+      // const response = await fetch("/challanges.json");
+      if (!response.ok) {
+        throw new Error(`HTTP error! status: ${response.status}`);
+      }
+
+      return response.json();
+    }
   },
   async deleteChallenge(id: string) {
-    const response = await fetch(
-      `${API_BASE_URL}${API_ENDPOINTS.CHALLENGE.DELETE(id)}`,
-      {
-        method: "DELETE",
-        credentials: "include", // JWT Auth
-        headers: {
-          "Content-Type": "application/json",
-        },
+
+    if (IS_LOCAL && IS_DEV) {
+      // Mock response für Testzwecke
+      const response = {
+        ok: true, // API-Antwort simulieren
+        status: 200,
+        statusText: "OK",
+        json: async () => ({ detail: "lol" }), // Dummy JSON-Antwort
+      };
+
+      const data = await response.json();
+
+      if (!response.ok) {
+        throw new Error(
+          data.detail || `Failed to delete challenge: ${response.status}`
+        );
       }
-    );
-    // Mock response für Testzwecke
-    // const response = {
-    //   ok: true, // API-Antwort simulieren
-    //   status: 200,
-    //   statusText: "OK",
-    //   json: async () => ({ detail: "lol" }), // Dummy JSON-Antwort
-    // };
 
-    const data = await response.json();
+      return data;
+    } else {
 
-    if (!response.ok) {
-      throw new Error(
-        data.detail || `Failed to delete challenge: ${response.status}`
+      const response = await fetch(
+        `${API_BASE_URL}${API_ENDPOINTS.CHALLENGE.DELETE(id)}`,
+        {
+          method: "DELETE",
+          credentials: "include", // JWT Auth
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
       );
-    }
 
-    return data;
+      const data = await response.json();
+
+      if (!response.ok) {
+        throw new Error(
+          data.detail || `Failed to delete challenge: ${response.status}`
+        );
+      }
+
+      return data;
+    }
   },
 };
